+++ dex/fuzz/src/fuzz_pair.rs
@@ -12,9 +12,8 @@ pub mod fuzz_pair_test {
     use rand::prelude::*;
 
     use crate::fuzz_data::fuzz_data_tests::*;
-    use pair::{
-        pair_actions::{add_liq::AddLiquidityModule, remove_liq::RemoveLiquidityModule},
-        *,
+    use pair::pair_actions::{
+        add_liq::AddLiquidityModule, remove_liq::RemoveLiquidityModule, swap::SwapModule,
     };
 
     pub fn add_liquidity<PairObjBuilder, FarmObjBuilder, FactoryObjBuilder, PriceDiscObjBuilder>(
