+++ dex/pair/src/pair_actions/common_methods.rs
@@ -8,4 +8,9 @@ pub trait CommonMethodsModule {
     fn is_state_active(&self, state: State) -> bool {
         state == State::Active || state == State::PartialActive
     }
+
+    #[inline]
+    fn can_swap(&self, state: State) -> bool {
+        state == State::Active
+    }
 }
