+++ dex/pair/src/pair_actions/mod.rs
@@ -3,3 +3,4 @@ pub mod common_methods;
 pub mod common_result_types;
 pub mod initial_liq;
 pub mod remove_liq;
+pub mod swap;
