+++ dex/pair/tests/pair_rs_test.rs
@@ -13,8 +13,10 @@ use multiversx_sc_scenario::{
     managed_address, managed_biguint, managed_token_id, managed_token_id_wrapped, rust_biguint,
     whitebox_legacy::TxTokenTransfer, DebugApi,
 };
-// use pair::safe_price::MAX_OBSERVATIONS;
-use pair::{config::MAX_PERCENTAGE, fee::FeeModule, locking_wrapper::LockingWrapperModule, Pair};
+use pair::{
+    config::MAX_PERCENTAGE, fee::FeeModule, locking_wrapper::LockingWrapperModule,
+    pair_actions::swap::SwapModule,
+};
 use pair_setup::*;
 use simple_lock::{
     locked_token::{LockedTokenAttributes, LockedTokenModule},
